feat: reorganize oso_agent
* Segregate out all tools as separate functions to be reused
* Create an agent registry, with both react and sql agents
* Remove system prompt as a config variable. We'll get this from Phoenix
