+++ warehouse/oso_agent/.env.example
@@ -12,4 +12,5 @@ AGENT_DATASET_NAME=
 AGENT_LLM__TYPE=local
 AGENT_OLLAMA_MODEL=llama3.2:3b
 #AGENT_LLM__TYPE="google_gemini"
+#AGENT_LLM__MODEL="models/gemini-2.0-flash"
 #AGENT_LLM__GOOGLE_API_KEY=
