+++ warehouse/oso_agent/oso_agent/agent/__init__.py
@@ -1,11 +1,9 @@
-from .agent import Agent
-from .config import AgentConfig
-from .errors import AgentConfigError, AgentError, AgentRuntimeError
+from .react_agent import create_react_agent
+from .registry import AgentRegistry
+from .sql_agent import create_sql_agent
 
 __all__ = [
-    "Agent",
-    "AgentConfig",
-    "AgentError",
-    "AgentConfigError",
-    "AgentRuntimeError",
+    "AgentRegistry",
+    "create_react_agent",
+    "create_sql_agent",
 ]
