+++ warehouse/oso_agent/oso_agent/cli/utils.py
@@ -1,16 +1,15 @@
 import click
 
-from ..agent.config import AgentConfig
+from ..util.config import AgentConfig
 
 pass_config = click.make_pass_decorator(AgentConfig, ensure=True)
 
-
 def common_options(f):
     """Common options decorator for commands that use the agent."""
-    f = click.option(
-        "--system-prompt",
-        "-p",
-        help="System prompt for the agent",
-        envvar="AGENT_SYSTEM_PROMPT",
-    )(f)
+#    f = click.option(
+#        "--system-prompt",
+#        "-p",
+#        help="System prompt for the agent",
+#        envvar="AGENT_SYSTEM_PROMPT",
+#    )(f)
     return f
