+++ warehouse/oso_agent/oso_agent/util/__init__.py
@@ -1,0 +1,24 @@
+"""Utility functions for the application."""
+
+from ..util.log import setup_logging
+from .config import (
+    AgentConfig,
+    GeminiLLMConfig,
+    GoogleGenAILLMConfig,
+    LocalLLMConfig,
+    agent_config_dict,
+)
+from .errors import AgentConfigError, AgentError, AgentMissingError, AgentRuntimeError
+
+__all__ = [
+    "setup_logging",
+    "agent_config_dict",
+    "AgentConfig",
+    "LocalLLMConfig",
+    "GeminiLLMConfig",
+    "GoogleGenAILLMConfig",
+    "AgentConfigError",
+    "AgentError",
+    "AgentRuntimeError",
+    "AgentMissingError",
+]
