--- warehouse/oso_agent/oso_agent/utils/__init__.py
@@ -1,6 +1,0 @@
-"""Utility functions for the application."""
-
-from ..utils.log import setup_logging
-from .config import agent_config_dict
-
-__all__ = ["setup_logging", "agent_config_dict"]
